requirements.txt
langchain==0.0.225
ctransformers==0.2.5
sentence-transformers==2.2.2
langchain-pinecone
langchain-community
pinecone-client
python-dotenv
flask
pypdf
-e .